Revert "Add Ironclad user editing."
This reverts commit 7c76c57e5580845e044baf63637ddd66c5a9c7be.
