+++ src/Lykke.Service.OAuth.Services/TokenService.cs
@@ -16,7 +16,7 @@ namespace Lykke.Service.OAuth.Services
     public class TokenService : ITokenService
     {
         private readonly IDatabase _redisDatabase;
-        private const string RedisPrefixIroncladRefreshTokens = "OAuth:ExternalRefreshTokens";
+        private const string RedisPrefixExternalRefreshTokens = "OAuth:ExternalRefreshTokens";
         private static readonly TimeSpan RefreshTokenWhitelistLifetime = TimeSpan.FromDays(30);
         private static readonly TimeSpan IroncladRefreshTokenLifetime = TimeSpan.FromDays(30);
         private readonly IHttpClientFactory _httpClientFactory;
@@ -92,7 +92,7 @@ namespace Lykke.Service.OAuth.Services
             if (string.IsNullOrWhiteSpace(refreshToken))
                 throw new ArgumentNullException(nameof(refreshToken));
 
-            var redisKey = GetIroncladRefreshTokensRedisKey(lykkeToken);
+            var redisKey = GetExternalRefreshTokensRedisKey(lykkeToken);
 
             return _redisDatabase.StringSetAsync(redisKey, refreshToken, IroncladRefreshTokenLifetime);
         }
@@ -103,7 +103,7 @@ namespace Lykke.Service.OAuth.Services
             if (string.IsNullOrWhiteSpace(lykkeToken))
                 throw new ArgumentNullException(nameof(lykkeToken));         
 
-            var redisKey = GetIroncladRefreshTokensRedisKey(lykkeToken);
+            var redisKey = GetExternalRefreshTokensRedisKey(lykkeToken);
 
             var ironcladRefreshToken = await _redisDatabase.StringGetAsync(redisKey);
 
@@ -116,6 +116,7 @@ namespace Lykke.Service.OAuth.Services
         /// <inheritdoc />
         public async Task<string> GetIroncladAccessTokenAsync(string lykkeToken)
         {
+
             if (string.IsNullOrWhiteSpace(lykkeToken))
                 throw new ArgumentNullException(nameof(lykkeToken));
 
@@ -131,10 +132,11 @@ namespace Lykke.Service.OAuth.Services
                 RefreshToken = ironcladRefreshToken
             });
 
-            if (tokenResponse.IsError)
-            {
-                throw new TokenNotFoundException(tokenResponse.Error);
-            }
+            //TODO:@gafanasiev Add error handling.
+            //if (tokenResponse.IsError)
+            //{
+            //    throw 
+            //}
 
             await SaveIroncladRefreshTokenAsync(lykkeToken, tokenResponse.RefreshToken);
 
@@ -160,12 +162,12 @@ namespace Lykke.Service.OAuth.Services
             }
         }
 
-        private static string GetIroncladRefreshTokensRedisKey(string lykkeToken)
+        private static string GetExternalRefreshTokensRedisKey(string lykkeToken)
         {
             if (string.IsNullOrWhiteSpace(lykkeToken))
                 throw new ArgumentNullException(nameof(lykkeToken));      
             
-            return $"{RedisPrefixIroncladRefreshTokens}:{lykkeToken}";
+            return $"{RedisPrefixExternalRefreshTokens}:{lykkeToken}";
         }
     }
 }
