+++ src/Lykke.Service.OAuth/Controllers/TokenController.cs
@@ -2,7 +2,6 @@
 using AspNet.Security.OpenIdConnect.Server;
 using Common.Log;
 using Core.Extensions;
-using Core.ExternalProvider;
 using Core.Services;
 using Lykke.Common.Log;
 using Lykke.Service.Session.Client;
@@ -11,33 +10,21 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Lykke.Service.OAuth.Controllers
 {
-    [Route("api/[controller]")]
     public class TokenController : Controller
     {
         private readonly ILog _log;
         private readonly IClientSessionsClient _clientSessionsClient;
         private readonly ITokenService _tokenService;
-        private readonly IIroncladService _ironcladService;
 
         public TokenController(
             ILogFactory logFactory,
             IClientSessionsClient clientSessionsClient,
-            ITokenService tokenService, 
-            IIroncladService ironcladService)
+            ITokenService tokenService)
 
         {
             _log = logFactory.CreateLog(this);
             _clientSessionsClient = clientSessionsClient;
             _tokenService = tokenService;
-            _ironcladService = ironcladService;
-        }
-
-        //TODO@gafanasiev Remove.
-        [HttpGet("~/testironclad")]
-        public async Task<IActionResult> TestIronclad()
-        {
-            await _ironcladService.AddClaim("a789e8b658624329a7053187fc5de5b6", "lsub", "test_id");
-            return Ok();
         }
 
         [HttpGet("~/getkyctoken")]
